+++ Sources/Engine/HTTP/Client/HTTPClient.swift
@@ -1,3 +1,66 @@
+/*
+     Specific sequences are:
+
+     CONNECTION ESTABLISHMENT
+
+     S: 220
+     E: 554
+
+     EHLO or HELO
+
+     S: 250
+     E: 504 (a conforming implementation could return this code only
+     in fairly obscure cases), 550, 502 (permitted only with an old-
+     style server that does not support EHLO)
+
+     MAIL
+
+     S: 250
+     E: 552, 451, 452, 550, 553, 503, 455, 555
+
+     RCPT
+
+     S: 250, 251 (but see Section 3.4 for discussion of 251 and 551)
+     E: 550, 551, 552, 553, 450, 451, 452, 503, 455, 555
+
+     DATA
+
+     I: 354 -> data -> S: 250
+
+     E: 552, 554, 451, 452
+
+     E: 450, 550 (rejections for policy reasons)
+
+     E: 503, 554
+
+     RSET
+
+     S: 250
+
+     VRFY
+
+     S: 250, 251, 252
+     E: 550, 551, 553, 502, 504
+
+     EXPN
+
+     S: 250, 252
+     E: 550, 500, 502, 504
+
+     HELP
+
+     S: 211, 214
+     E: 502, 504
+
+     NOOP
+
+     S: 250
+     
+     QUIT
+     
+     S: 221
+*/
+
 public enum HTTPClientError: ErrorProtocol {
     case invalidRequestHost
     case invalidRequestScheme
