+++ Sources/SMTP/EmailAddress+StringLiteralConvertible.swift
@@ -1,0 +1,13 @@
+extension EmailAddress: StringLiteralConvertible {
+    public init(stringLiteral string: String) {
+        self.init(name: nil, address: string)
+    }
+
+    public init(extendedGraphemeClusterLiteral string: String){
+        self.init(name: nil, address: string)
+    }
+
+    public init(unicodeScalarLiteral string: String){
+        self.init(name: nil, address: string)
+    }
+}
