+++ Sources/SMTP/EmailAddress.swift
@@ -1,0 +1,9 @@
+public struct EmailAddress {
+    public let name: String?
+    public let address: String
+
+    public init(name: String? = nil, address: String) {
+        self.name = name
+        self.address = address
+    }
+}
