+++ Sources/SMTP/SMTPAuthMethod.swift
@@ -1,0 +1,5 @@
+internal enum SMTPAuthMethod {
+    case plain
+    case login
+    // TODO: Support additional auth methods
+}
