+++ Sources/SMTP/SMTPClient+Transmit.swift
@@ -1,4 +1,5 @@
 import Base
+import Engine
 
 private let crlf: Bytes = [.carriageReturn, .newLine]
