+++ Sources/SMTP/SMTPClientError.swift
@@ -1,0 +1,6 @@
+public enum SMTPClientError: ErrorProtocol {
+    case initializationFailed(code: Int, greeting: String)
+    case initializationFailed554(reason: String)
+    case invalidMultilineReplyCode(expected: Int, got: Int)
+    case invalidEhloHeader
+}
