+++ Sources/SMTP/SMTPCredential.swift
@@ -1,0 +1,9 @@
+public struct SMTPCredentials {
+    public let user: String
+    public let pass: String
+
+    public init(user: String, pass: String) {
+        self.user = user
+        self.pass = pass
+    }
+}
